pyproject.toml
[project]
name = "pyspark-transformations"
version = "0.1.0"
description = "Project processing data from CSV files using PySpark, applying transformations, and saving the result to output CSV file"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyspark>=3.5.0",
    "python-dotenv>=1.1.1",
    "pytest>=8.4.1",
    "requests>=2.32.4",
]
